+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 CHANGELOG
 ==========
 
+- 2026-05-22 wsEventEmitter: Add /top_laser_3d/scan event emitter
+- 2026-05-22 robotApi: Add calibrateTopLidar service API
 - 2026-05-21 robotApi: Add calibrateHeadLidar and calibrateRearLidar service APIs
 - 2026-05-11 submapCache: Add generic submap cache
 - 2026-05-11 robotApi: Add SubmapQueryV2 fetch API